+++ scripts/weekly/S4DownloadClassSchedulesBeta.php
@@ -36,10 +36,20 @@ class S4DownloadClassSchedulesBeta implements iSubscript {
   }
   
   public function downloadClassSchedule($sCourseCode) {
-    echo ".";
-    $sFile = dirname(__FILE__) . "/../../cache/class_beta/" . $sCourseCode . ".ics";
-    //http://croosters.saxion.nl/ical/group/EIN1Va.ics
-    file_put_contents($sFile, fopen("http://croosters.saxion.nl/ical/group/" . $sCourseCode . ".ics", 'r'));
+$sFile = dirname(__FILE__) . "/../../cache/class_beta/" . $sCourseCode . ".ics";
+    if (file_exists($sFile)) {
+      if (filemtime($sFile)  >= time() - 43200) {
+        echo "_";
+      } else {
+        //http://croosters.saxion.nl/ical/group/EIN1Va.ics
+        file_put_contents($sFile, fopen("http://croosters.saxion.nl/ical/group/" . $sCourseCode . ".ics", 'r'));
+        echo "U";
+      }
+    } else {
+      //http://croosters.saxion.nl/ical/group/EIN1Va.ics
+      file_put_contents($sFile, fopen("http://croosters.saxion.nl/ical/group/" . $sCourseCode . ".ics", 'r'));
+      echo "N";
+    }
   }
   
 }
